Move authenticators to auth package in github
