+++ enterprise/cmd/frontend/internal/app/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//internal/database",
         "//internal/extsvc",
         "//internal/extsvc/github",
+        "//internal/extsvc/github/auth",
         "//internal/jsonc",
         "//internal/observation",
         "//internal/types",
