+++ enterprise/cmd/frontend/internal/app/app.go
@@ -26,6 +26,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/github"
+	ghauth "github.com/sourcegraph/sourcegraph/internal/extsvc/github/auth"
 	"github.com/sourcegraph/sourcegraph/internal/jsonc"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/types"
@@ -55,7 +56,7 @@ func Init(
 		return nil
 	}
 
-	auther, err := github.NewGitHubAppAuthenticator(config.AppID, config.PrivateKey)
+	auther, err := ghauth.NewGitHubAppAuthenticator(config.AppID, config.PrivateKey)
 	if err != nil {
 		return errors.Wrap(err, "new authenticator with GitHub App")
 	}
