+++ enterprise/internal/authz/github/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//internal/extsvc",
         "//internal/extsvc/auth",
         "//internal/extsvc/github",
+        "//internal/extsvc/github/auth",
         "//internal/rcache",
         "//internal/types",
         "//lib/errors",
