+++ internal/extsvc/github/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "github",
     srcs = [
-        "authenticators.go",
         "common.go",
         "doc.go",
         "v3.go",
@@ -25,7 +24,6 @@ go_library(
         "//internal/ratelimit",
         "//internal/trace/ot",
         "//lib/errors",
-        "@com_github_golang_jwt_jwt_v4//:jwt",
         "@com_github_google_go_github//github",
         "@com_github_google_go_github_v41//github",
         "@com_github_graphql_go_graphql//language/ast",
@@ -42,7 +40,6 @@ go_test(
     name = "github_test",
     timeout = "short",
     srcs = [
-        "authenticators_test.go",
         "common_test.go",
         "v3_test.go",
         "v4_test.go",
