+++ internal/extsvc/github/auth/authenticators.go
@@ -1,4 +1,4 @@
-package github
+package auth
 
 import (
 	"context"
